shared wallet forward get
